Route all three Live Activity start paths through a shared prominence helper
